+++ lib/config/config_default.ex
@@ -19,4 +19,5 @@ defmodule FormFlow.Config.Default do
   defdelegate enabled_flow_types(context, config_data), to: Config.Default
   defdelegate enabled_form_types(context, config_data), to: Config.Default
   defdelegate handle_mount(context, config_data), to: Config.Default
+  defdelegate flow_instances_query(context, config_data), to: Config.Default
 end